fix/ supabase 를 이용하는 코드를 구조에 맞게 이동했습니다. #3
