+++ src/app/(home)/components/home-diary.tsx
@@ -1,10 +1,8 @@
 import React from "react";
-import {
-  checkEmotion,
-  getPostEmotionByUserId,
-} from "../../../utils/home-emotion.util";
+import { checkEmotion } from "../../../utils/home-emotion.util";
 import { HomePost } from "../types/HomePost";
 import Image from "next/image";
+import { getPostEmotionByUserId } from "@/services/home-services";
 
 const HomeDiary = async () => {
   const posts: HomePost[] = await getPostEmotionByUserId();
